do_classroom/requirements.txt
appnope==0.1.0
astroid==2.3.3
backcall==0.1.0
certifi==2019.11.28
cffi==1.13.2
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
cryptography==2.8
decorator==4.4.0
defusedxml==0.6.0
Django==2.2.7
django-cors-headers==3.2.0
django-webpack-loader==0.6.0
djangorestframework==3.11.0
drf-yasg==1.17.0
ecdsa==0.14.1
gunicorn==19.9.0
idna==2.8
inflection==0.3.1
ipython==7.6.1
ipython-genutils==0.2.0
isort==4.3.21
itypes==1.1.0
jedi==0.14.0
Jinja2==2.10.3
jsonpickle==1.2
lazy-object-proxy==1.4.3
MarkupSafe==1.1.1
mccabe==0.6.1
oauthlib==3.1.0
packaging==19.2
parso==0.5.0
pbr==5.4.0
pexpect==4.7.0
pickleshare==0.7.5
prompt-toolkit==2.0.9
protobuf==3.7.1
ptyprocess==0.6.0
pycparser==2.19
Pygments==2.4.2
PyJWT==1.7.1
pylint==2.4.3
pylint-django==2.0.13
pylint-plugin-utils==0.6
pyparsing==2.4.5
python-digitalocean==1.14.0
python-social-auth==0.3.6
python3-openid==3.1.0
pytz==2019.3
requests==2.22.0
requests-oauthlib==1.3.0
ruamel.yaml==0.16.5
ruamel.yaml.clib==0.2.0
six==1.13.0
social-auth-core==3.2.0
sqlparse==0.3.0
sshpubkeys==3.1.0
stevedore==1.30.1
traitlets==4.3.2
typed-ast==1.4.0
uritemplate==3.0.0
urllib3==1.25.7
virtualenv==16.6.1
virtualenv-clone==0.5.3
virtualenvwrapper==4.8.4
wcwidth==0.1.7
wrapt==1.11.2
xkcdpass==1.17.3
